+++ src/config.rs
@@ -123,6 +123,10 @@ pub struct TestConfig {
     #[serde(skip)]
     pub overwrite_tests: bool,
 
+    /// Interactively review each failing test file, accepting or rejecting its diff one by one.
+    #[serde(skip)]
+    pub interactive: bool,
+
     /// Arguments to always include in the command-line args for testing the program.
     /// For example, if this is `foo` and the test specifies `args: bar baz` then the
     /// binary will be invoked via `<binary> foo bar baz <filename> <args-after>`
@@ -259,6 +263,7 @@ impl TestConfig {
             test_exit_status_prefix: prefixed(test_exit_status_prefix),
             test_line_prefix,
             overwrite_tests,
+            interactive: false,
             base_args: String::new(),
             base_args_after: String::new(),
         }