+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.4.3
+
+- Added `--interactive`/`-i` flag to review failing test files one by one
+
 # 1.4.2
 
 - Added `glob` option to `goldentests.toml`
@@ -5,4 +9,4 @@
 
 # 1.4.1
 
-- Added `-j<N>`/`--jobs <N>` argument to optionally specify the number of parallel jobs if the `parallel` feature is used.
+- Added `-j<N>`/`--jobs <N>` argument to optionally specify the number of parallel jobs if the `parallel` feature is used